+++ regionmask/tests/test_mask_defined_regions.py
@@ -39,6 +39,9 @@ def _test_mask_equal_defined_regions(region, ds, mask_method):
 @pytest.mark.parametrize("defined_region", REGIONS, ids=str)
 @pytest.mark.parametrize("ds", DATASETS)
 def test_mask_equal_defined_regions(defined_region, ds):
+    if defined_region.skip_mask_test:
+        pytest.skip(reason=f"Manally skipping {defined_region.region_name}")
+
     # a loop over DATASETS is not faster - due to caching of the regions
     region = attrgetter(defined_region.region_name)(defined_regions)
     mask_method = "mask_3D" if defined_region.overlap else "mask"