+++ regionmask/tests/utils.py
@@ -88,6 +88,7 @@ class DefinedRegion:
     region_name: str
     n_regions: int
     overlap: bool = False
+    skip_mask_test: bool = False
 
     def __str__(self):
         # used as name (`ids`) for parametrize
@@ -106,6 +107,7 @@ REGIONS = [
 _REGIONS_NATURAL_EARTH_v4_1_0 = [
     DefinedRegion("natural_earth_v4_1_0.countries_110", 177),
     DefinedRegion("natural_earth_v4_1_0.countries_50", 241),
+    DefinedRegion("natural_earth_v4_1_0.countries_10", 258),
     DefinedRegion("natural_earth_v4_1_0.us_states_50", 51),
     DefinedRegion("natural_earth_v4_1_0.us_states_10", 51),
     DefinedRegion("natural_earth_v4_1_0.land_110", 1),
@@ -117,6 +119,7 @@ _REGIONS_NATURAL_EARTH_v4_1_0 = [
 _REGIONS_NATURAL_EARTH_v5_0_0 = [
     DefinedRegion("natural_earth_v5_0_0.countries_110", 177),
     DefinedRegion("natural_earth_v5_0_0.countries_50", 242),
+    DefinedRegion("natural_earth_v5_0_0.countries_10", 258, skip_mask_test=True),
     DefinedRegion("natural_earth_v5_0_0.us_states_50", 51),
     DefinedRegion("natural_earth_v5_0_0.us_states_10", 51),
     DefinedRegion("natural_earth_v5_0_0.land_110", 1),